LayoutTests/imported/w3c/web-platform-tests/navigation-api/ordering-and-transition/transition-finished-mark-as-handled-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: oh no!

Harness Error (FAIL), message = Unhandled rejection: oh no!

FAIL navigation.transition.finished must not trigger unhandled rejections null is not an object (evaluating 'navigation.transition.finished')
